pyproject.toml
[tool.poetry]
name = "kuadrant-testsuite"
version = "0.1.0"
description = "Tests for Kuadrant project"
authors = ["Kuadrant QE <kuadrantdev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.group.main.dependencies]
python = "^3.11"
typing_extensions = "*"
pytest-xdist = "*"
pytest = "*"
pytest-asyncio = ">=0.23.4"
pytest-html = "*"
dynaconf = "*"
python-keycloak = ">=3.6"
python-jose = "*"
lxml = "*"
cryptography = "*"
backoff = "*"
httpx = { version = "*", extras = ["http2"] }
openshift-client = ">=2"
apyproxy = "*"
weakget = "*"
pyyaml = "*"


[tool.poetry.group.dev.dependencies]
mypy = "*"
pylint = "*"
types-PyYAML = "*"
types-python-jose = "*"
black = {version = "*", extras = ["d"]}


# Black
[tool.black]
line-length = 120

# Pytest:
[tool.pytest.ini_options]
markers = [
    "issue: Reference to covered issue",
    "performance: Performance tests have unique needs",
    "authorino: Test is using Authorino features",
    "standalone_only: Test is using features available only in standalone mode, without Kuadrant",
    "kuadrant_only: Test is using features available only in Kuadrant mode",
    "limitador: Test is using Limitador features",
]
filterwarnings = [
    "ignore: WARNING the new order is not taken into account:UserWarning",
    "ignore::urllib3.exceptions.InsecureRequestWarning",
    "ignore::cryptography.utils.CryptographyDeprecationWarning"
]
log_format = "%(asctime)s %(levelname)s:%(name)s:%(message)s"
log_date_format = "%H:%M:%S %z"
log_level = "INFO"
junit_logging = "all"
junit_family = "xunit2"
asyncio_mode = "auto"
xfail_strict = true

# Pylint:
[tool.pylint.FORMAT]
max-line-length = 120
disable = [
    "duplicate-code", # reports false alarms AND can't be disabled locally; pylint issue #214
    "fixme", # ignore TODOs
    "redefined-outer-name",
    "too-many-arguments",
    "too-few-public-methods"]

[tool.pylint.BASIC]
good-names=["i","j","k",
           "pytestmark",
           "logger",
           "ca", "gw"]

# Mypy:
[tool.mypy]
implicit_optional = true

[[tool.mypy.overrides]]
module = ["dynaconf.*", "keycloak.*", "weakget.*", "openshift_client.*", "apyproxy.*", "click.*", "py.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
